requirements.txt
fastapi
uvicorn
pytest
llama-cpp-python
pyyaml
